[BUGFIX] Compatibility fix for HDF5 1.10

Approved-by: dcollins4096 <dcollins4096@gmail.com>
Approved-by: John  Regan <johnanthonyregan@gmail.com>
Approved-by: Brian O'shea <oshea@msu.edu>

--HG--
branch : week-of-code

+++ src/enzo/macros_and_parameters.h
@@ -152,7 +152,7 @@ typedef long long int  Elong_int;
 
 typedef int            MPI_Arg;
 
-typedef int            HDF5_hid_t;
+typedef long long int   HDF5_hid_t;
 
 /* HDF5 definitions */
